+++ mdserver/config.py
@@ -10,6 +10,12 @@
 # or directories to load specified in the `[_include_]` section of the main
 # config file.
 #
+# Note: Bottle's ConfigDict class uses the ExtendedInterpolation class when
+# creating its ConfigParser objects - we override that, disabling
+# interpolation entirely. This ensures that configuration values can include
+# the $ and { characters without needing to be escaped. This avoids issues
+# with password hashes in particular, as well as with other user-specified
+# free-form strings.
 
 import logging
 import os
@@ -121,7 +127,7 @@ def load_dir(app, dirname):
     abs_files = [os.path.join(dirname, fname) for fname in files]
     for f in abs_files:
         early_logger.debug("Loading config from %s", f)
-        app.config.load_config(f)
+        app.config.load_config(f, interpolation=None)
     app.config["_files_." + dirname] = ",".join(files)
 
 
@@ -129,14 +135,15 @@ def load_files(app, files):
     """Load the contents of the specified files."""
     for f in files:
         early_logger.debug("Loading config from %s", f)
-        app.config.load_config(f)
+        app.config.load_config(f, interpolation=None)
     app.config["_files_.files"] = ",".join(files)
 
 
 def load(app, filename):
     """Load configuration into app.config from the specified file.
 
-    File is a standard ini style config file.
+    File is a standard ini style config file. Files are parsed using the
+    standard ConfigParser class, with interpolation=None.
 
     Additional files and/or directories can be included via the `_include_`
     section, with a colon-separated list of files or paths specified via the
@@ -151,7 +158,7 @@ def load(app, filename):
     conf_files = []
     if os.path.exists(filename):
         early_logger.debug("Loading config from %s", filename)
-        app.config.load_config(filename)
+        app.config.load_config(filename, interpolation=None)
         app.config["_files_.main"] = filename
         if "_include_.directories" in app.config:
             dirlist = app.config["_include_.directories"]